requirements.txt
appnope==0.1.2
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
DateTime==4.3
decorator==4.4.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
jsonlines==2.0.0
jsonpointer==2.0
jsonschema==3.2.0
lxml==4.6.2
MarkupSafe==1.1.1
packaging==20.8
parsel==1.6.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.10
Protego==0.1.16
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.4
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
python-decouple==3.4
python-slugify==4.0.1
pytz==2020.5
queuelib==1.5.0
requests==2.25.1
rfc3987==1.3.8
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
spidermon==1.14.0
strict-rfc3339==0.7
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
Twisted==20.3.0
urllib3==1.26.2
uuid==1.30
w3lib==1.22.0
wcwidth==0.2.5
webcolors==1.11.1
zope.interface==5.2.0
